Swagger: sorting tags groups & documenting the genre's API
